fix: simplify ranking metric + fix Прощёное воскресенье spelling
Ranking: score = correct_answers × 10 (честная накопительная метрика).
8/10 = 80 очков, 16/20 = 160 очков. Убраны полоса точности, количество
попыток и легенда — только имя и очки.

TeacherDashboard: ruTotalScore теперь суммирует details.correct × 10
вместо поля score (которое было процентом, не зависящим от объёма).

SpellingNN: исправлена ошибка — «Прощёное воскресенье» писалось как double
(НН), тогда как по нормам русского языка это устойчивое выражение с одной Н.

+++ src/exercises/SpellingNN.tsx
@@ -71,7 +71,7 @@ const ITEMS: Item[] = [
   { display: 'чека_ый', single: 'чеканый', double: 'чеканный', answer: 'double' },
   { display: 'жела_ый', single: 'желаный', double: 'желанный', answer: 'double' },
   { display: 'жема_ый', single: 'жеманый', double: 'жеманный', answer: 'double' },
-  { display: 'Прощё_ое воскресенье', single: 'Прощёное воскресенье', double: 'Прощённое воскресенье', answer: 'double' },
+  { display: 'Прощё_ое воскресенье', single: 'Прощёное воскресенье', double: 'Прощённое воскресенье', answer: 'single' },
 ];
 
 export default function SpellingNN({ onBack }: Props) {
